docs: add form example using NextJS & React apis
